FDA_agent/requirements.txt
openai>=1.0.0
numpy>=1.21.0
tenacity>=8.0.0
python-dotenv>=0.19.0
requests>=2.26.0
fastapi>=0.68.0
pydantic>=1.8.0
uvicorn>=0.15.0
backoff>=2.0.0
